refactor: share the list toolbar layout and mobile search hooks
