+++ frontend/src/pages/transactions/components/toolbar/types.ts
@@ -1,4 +1,3 @@
-import type { RefObject } from 'react'
 import type { Category } from '@/api/categories'
 import type { OptionItem } from '@/components/filters/OptionList'
 import type { TransactionListAccount, TransactionListFilters } from '@/pages/transactions/types/transactionList'
@@ -25,11 +24,4 @@ export type TransactionListToolbarProps = {
 export type TransactionToolbarOptions = {
   accountOptions: OptionItem[]
   categoryOptions: OptionItem[]
-}
-
-export type DesktopTransactionToolbarRefs = {
-  toolbarRef: RefObject<HTMLDivElement | null>
-  controlsRef: RefObject<HTMLDivElement | null>
-  filterGroupRef: RefObject<HTMLDivElement | null>
-  createMeasureRef: RefObject<HTMLButtonElement | null>
 }
